ch11-multi-agent/pyproject.toml
[project]
name = "ch11-multi-agent"
version = "0.1.0"
description = "Chapter 11: HR 전문팀 시스템 (Supervisor 패턴)"
requires-python = ">=3.11"
dependencies = [
    "langchain>=0.3.25",
    "langchain-core>=0.3.61",
    "langchain-openai>=0.3.18",
    "langgraph>=0.4.8",
    "pydantic>=2.11.5",
    "pydantic-settings>=2.9.1",
    "python-dotenv>=1.1.0",
    "streamlit>=1.45.1",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/hr_team"]
